fs_mgr: When formating ext4 volumes, respect the length parameter
When formatting a volume because it is marked as formattable, the
length parameter is discarded which makes fs_mgr write a filesystem
to the full length of the block device.

This patch adds length semantics to ext4 formatting, if the length
is greater than zero, use that, if it isn't subtract that size
from the block size.

Change-Id: I0eadade18396233f08d2eac6d7158a78a43f6b26
Ticket: CYNGNOS-931