healthd: Add fast charging UI support [2/3] (Update)
Exposes the charge_type battery property.

At the same time provide libhealthd hooks to customize the
processing of the status and charge_type strings.  These
strings are not standardized at this time when it comes to
fast charging.  For example,

* Asus uses status = "Quick charging" or status = "Not quick charging"
  when plugged into a fast charger to indicate that it is charging
  on a fast charger.

* Motorola uses charge_type sanely.

* Oppo uses a completely different boolean property.

and therefore we do not offer a default parsing of the charge
type and require that users add a libhealthd that defines the path
to read and the function to interpret the result of reading that
file.

Change-Id: Icaa1f29566f15bf4ce3a0705cef1c14eceea12fb